# swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: REST API
  description: ''
host: '192.168.99.105:3001'
basePath: /
tags: []
schemes:
  - http
securityDefinitions: {}
consumes: []
produces: []
paths:
  /candidates:
    get:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
  /profile:
    get:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /register:
    post:
      tags: []
      description: ''
      parameters: []
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
  /login:
    post:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
  /logout:
    post:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
  /profile/update:
    put:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /vote/add:
    post:
      tags: []
      description: ''
      parameters: []
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /vote/discard:
    post:
      tags: []
      description: ''
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
definitions: {}